feat: Add new sociodemographic properties
The SociodemographicDetails model has been updated to include new properties related to sociodemographic information. These properties include \hipertPulm\ for pulmonary hypertension, \centerAttention\ for the center of attention, \liveAlone\ to indicate if the person lives alone, \hasTechUseDifficulty\ to indicate if there is difficulty in using technology, and \needsCellphoneAssistance\ to indicate if there is a need for cellphone assistance. This update enhances the sociodemographic data collection capabilities of the application.